[abus] Add abus and a char device on it
Create a bus, "abus", and create a attribute, "device_action", for it.
When read this attribute under /sys/bus/abus/, it can emulate the device
insert and remove action. This device "abus_char_device" is a char
device, it belong to the bus, "abus", and it belong to the class,
"aclass", as design. It can be detected and its device file created
autumatically.

Issues:
1, When read "device_action" attribute, following error pops up when adding
device.
<3>[   80.930040] Device 'abus_char_device' does not have a release()
function, it is broken and must be fixed.

2, Have not found the right way to add the device "abus_char_device" to
the class "aclass".

next step:
Add a driver for this device and emulate the bus function match() and
probe().